+++ react-native.config.js
@@ -7,11 +7,4 @@ module.exports = {
       },
     },
   },
-  "react-native-video": {
-    platforms: {
-      android: {
-        sourceDir: "../node_modules/react-native-video/android-exoplayer",
-      },
-    },
-  },
 };