+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- We have upgraded `react-native-video` to 6.4.5
+
 - We have fixed OTA updates compatibility issues on iOS 18.
 
 ## [9.0.0] - 2024-07-12